Fix incorrect use of half for lighting attenuation.
Sync Graphics commit.
Lighting attenuation data is stored as float in shader. The Lighting function took it incorrectly as half which would cause light falloff clipping with large high intensity lights on mobile platforms. This only affected pixel light path while the vertex light path was correct.
The Lighting function now takes attenuation as float.

+++ Shaders/Fabric/Lighting.hlsl
@@ -40,7 +40,7 @@ void SheenScattering(BRDFData brdfData, half NoH, half NoV, half NoL, out half S
 #endif
 }
 
-half3 AnisoFabricLighting(BRDFData brdfData, half3 lightColor, half3 lightDirectionWS, half lightAttenuation,
+half3 AnisoFabricLighting(BRDFData brdfData, half3 lightColor, half3 lightDirectionWS, float lightAttenuation,
                           half3 normalWS, half3 viewDirectionWS, bool specularHighlightsOff,
                           AnisotropyData anisotropyData, SheenData SheenData, BRDFOcclusionFactor aoFactor)
 {
@@ -109,7 +109,7 @@ half3 AnisoFabricLighting(BRDFData brdfData, half3 lightColor, half3 lightDirect
     return brdf * Radiance;
 }
 
-half3 FabricLighting(BRDFData brdfData, half3 lightColor, half3 lightDirectionWS, half lightAttenuation,
+half3 FabricLighting(BRDFData brdfData, half3 lightColor, half3 lightDirectionWS, float lightAttenuation,
                     half3 normalWS, half3 viewDirectionWS,
                     bool specularHighlightsOff, SheenData SheenData, BRDFOcclusionFactor aoFactor)
 {
